refactor(live-preview): rename new/ to inline/ (canonical path)
Context
-------
Fase 12.5 step 3 of 5. With the legacy pipeline gone and the flag
removed, the "new/" naming is an accidental historical artifact —
there is no non-new counterpart anymore. Rename to "inline/" so the
directory describes its responsibility (the inline decoration
pipeline), not its relationship to a deleted predecessor.

Also dropping the `new-` prefix from the public entry file for the
same reason: `new-inline-extensions.ts` → `inline-extensions.ts`.
`new-pipeline-dom.test.ts` → `pipeline-dom.test.ts`.

JSDoc drift: every reference to "Phase N" or "new pipeline" in the
inline/ files is replaced with durable phrasing ("the inline
pipeline", "handlers under handlers/"). The Phase-numbered history
lives in docs/adr/0010 and tasks/todo/; the runtime code shouldn't
encode its own chronology.

Exported function name `newInlineExtensions()` kept as-is for now —
renaming cascades into live-preview.ts and every handler-adjacent
test. Low cost to change later if desired.

Solution
--------
git mv of the directory + the two specifically renamed files. All
relative imports inside inline/ keep working unchanged (they use
`./` and `../` that the rename doesn't break). External consumers
(live-preview.ts + tests) had their absolute + explicit paths sed-
rewritten:
  - live-preview/new/new-inline-extensions → live-preview/inline/inline-extensions
  - live-preview/new/inline-formatting-plugin → live-preview/inline/inline-formatting-plugin
  - live-preview/new/markdown-highlight-style → live-preview/inline/markdown-highlight-style
  - live-preview/new/handlers/... → live-preview/inline/handlers/...

JSDoc updates in the three files that carried "Phase N" labels:
inline-extensions.ts, inline-formatting-plugin.ts,
markdown-highlight-style.ts, handlers/inline-marks-handlers.ts.

Full suite: 5569 tests green, tsc clean.

Files moved (21 files + 1 test moved)
-------------------------------------
- src/lib/core/markdown-editor/extensions/live-preview/new/  →  …/inline/
  (12 files including handlers/)
- src/tests/lib/core/markdown-editor/extensions/live-preview/new/  →  …/inline/
  (14 files including handlers/)
- new/new-inline-extensions.ts → inline/inline-extensions.ts
- new/new-pipeline-dom.test.ts → inline/pipeline-dom.test.ts

Files modified (imports only)
-----------------------------
- src/lib/core/markdown-editor/extensions/live-preview/live-preview.ts
- src/tests/lib/core/markdown-editor/extensions/live-preview/live-preview.test.ts
- src/tests/lib/core/markdown-editor/extensions/live-preview/inline/*.test.ts
- src/tests/lib/core/markdown-editor/extensions/live-preview/inline/handlers/*.test.ts

JSDoc rewrites
--------------
- inline/inline-extensions.ts: "new inline pipeline" → "inline
  pipeline"; remove the experimental.newLivePreview reference.
- inline/markdown-highlight-style.ts: drop "(Phase 3)"/"(Phase 4)"
  breadcrumbs; list the five tag → class contracts.
- inline/inline-formatting-plugin.ts: drop "During Phase 2" comment.
- inline/handlers/inline-marks-handlers.ts: drop "(Phase 3)".
