+++ src/lib/core/markdown-editor/extensions/live-preview/inline/handlers/block-reference-handler.ts
