+++ src/tests/lib/core/markdown-editor/extensions/live-preview/inline/handlers/heading-handler.test.ts
@@ -6,8 +6,8 @@ import {
 	buildInlineDecorations,
 	registerInlineHandler,
 	_clearInlineHandlers,
-} from '$lib/core/markdown-editor/extensions/live-preview/new/inline-formatting-plugin';
-import { headingHandlers } from '$lib/core/markdown-editor/extensions/live-preview/new/handlers/heading-handler';
+} from '$lib/core/markdown-editor/extensions/live-preview/inline/inline-formatting-plugin';
+import { headingHandlers } from '$lib/core/markdown-editor/extensions/live-preview/inline/handlers/heading-handler';
 import { createMarkdownState } from '../../../../test-helpers';
 
 interface Spec {
