+++ src/tests/lib/core/markdown-editor/extensions/live-preview/inline/handlers/wikilink-handler.test.ts
@@ -6,8 +6,8 @@ import {
 	buildInlineDecorations,
 	registerLineHandler,
 	_clearInlineHandlers,
-} from '$lib/core/markdown-editor/extensions/live-preview/new/inline-formatting-plugin';
-import { wikilinkHandler } from '$lib/core/markdown-editor/extensions/live-preview/new/handlers/wikilink-handler';
+} from '$lib/core/markdown-editor/extensions/live-preview/inline/inline-formatting-plugin';
+import { wikilinkHandler } from '$lib/core/markdown-editor/extensions/live-preview/inline/handlers/wikilink-handler';
 import { createMarkdownState } from '../../../../test-helpers';
 
 function collect(decoSet: DecorationSet) {
